Update detective to latest version
Detective 2.x uses an old version of Esprima for parsing the files, this
unforunately does not support ES6 syntax (I specifically had problems
with template strings). The latest version of detective uses acorn which
does support template strings correctly.

Also updated the test to work in a later version of browserify as the
output from bundle and the place into which to pass the `debug`
parameter have moved.

(Also hi!)